+++ v3/operations.go
@@ -8565,7 +8565,7 @@ func (c Client) GetInstancePool(ctx context.Context, id UUID) (*InstancePool, er
 
 type UpdateInstancePoolRequest struct {
 	// Instance Pool Anti-affinity Groups
-	AntiAffinityGroups []AntiAffinityGroup `json:"anti-affinity-groups,omitempty"`
+	AntiAffinityGroups []AntiAffinityGroup `json:"anti-affinity-groups"`
 	// Deploy target
 	DeployTarget *DeployTarget `json:"deploy-target,omitempty"`
 	// Instance Pool description
@@ -8573,7 +8573,7 @@ type UpdateInstancePoolRequest struct {
 	// Instances disk size in GB
 	DiskSize int64 `json:"disk-size,omitempty" validate:"omitempty,gte=10,lte=50000"`
 	// Instances Elastic IPs
-	ElasticIPS []ElasticIP `json:"elastic-ips,omitempty"`
+	ElasticIPS []ElasticIP `json:"elastic-ips"`
 	// Prefix to apply to Instances names (default: pool)
 	InstancePrefix *string `json:"instance-prefix,omitempty"`
 	// Compute instance type
@@ -8586,14 +8586,14 @@ type UpdateInstancePoolRequest struct {
 	// Instance Pool name
 	Name string `json:"name,omitempty" validate:"omitempty,gte=1,lte=255"`
 	// Instance Pool Private Networks
-	PrivateNetworks    []PrivateNetwork   `json:"private-networks,omitempty"`
+	PrivateNetworks    []PrivateNetwork   `json:"private-networks"`
 	PublicIPAssignment PublicIPAssignment `json:"public-ip-assignment,omitempty"`
 	// Instance Pool Security Groups
-	SecurityGroups []SecurityGroup `json:"security-groups,omitempty"`
+	SecurityGroups []SecurityGroup `json:"security-groups"`
 	// SSH key
 	SSHKey *SSHKey `json:"ssh-key,omitempty"`
 	// Instances SSH keys
-	SSHKeys []SSHKey `json:"ssh-keys,omitempty"`
+	SSHKeys []SSHKey `json:"ssh-keys"`
 	// Instance template
 	Template *Template `json:"template,omitempty"`
 	// Instances Cloud-init user-data